Register artifacts and web3 as globals for scripts
