+++ deploy/logger.js
@@ -1,0 +1,5 @@
+module.exports = function() {
+  if (global.log) {
+    console.log.apply(this, arguments);
+  }
+}
